pyproject.toml
[tool.poetry]
name = "bibliochouette"
version = "0.1.0"
description = "Library management"
authors = ["Jérôme Desseaux"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
djangorestframework-jwt = "^1.11.0"
drf-spectacular = "^0.25.1"
python-dotenv = "^0.21.1"
djangorestframework-simplejwt = "^5.2.2"
django-rest-passwordreset = "^1.3.0"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
